# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "MacroGoalsRequest": {
        "properties": {
          "calories_target": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Calories Target",
            "type": "integer"
          },
          "carbs_target": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Carbs Target",
            "type": "integer"
          },
          "fat_target": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Fat Target",
            "type": "integer"
          },
          "protein_target": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Protein Target",
            "type": "integer"
          }
        },
        "title": "MacroGoalsRequest",
        "type": "object"
      },
      "MacroGoalsResponse": {
        "properties": {
          "calories_target": {
            "default": 0,
            "title": "Calories Target",
            "type": "integer"
          },
          "carbs_target": {
            "default": 0,
            "title": "Carbs Target",
            "type": "integer"
          },
          "fat_target": {
            "default": 0,
            "title": "Fat Target",
            "type": "integer"
          },
          "protein_target": {
            "default": 0,
            "title": "Protein Target",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "calories_target",
          "protein_target",
          "carbs_target",
          "fat_target"
        ],
        "title": "MacroGoalsResponse",
        "type": "object"
      },
      "MealItemRequest": {
        "properties": {
          "calories": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Calories",
            "type": "integer"
          },
          "carbs": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Carbs",
            "type": "integer"
          },
          "fat": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Fat",
            "type": "integer"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "protein": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Protein",
            "type": "integer"
          }
        },
        "required": [
          "name"
        ],
        "title": "MealItemRequest",
        "type": "object"
      },
      "MealItemResponse": {
        "properties": {
          "calories": {
            "default": 0,
            "title": "Calories",
            "type": "integer"
          },
          "carbs": {
            "default": 0,
            "title": "Carbs",
            "type": "integer"
          },
          "fat": {
            "default": 0,
            "title": "Fat",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "meal_id": {
            "title": "Meal Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "protein": {
            "default": 0,
            "title": "Protein",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "meal_id",
          "name",
          "calories",
          "protein",
          "carbs",
          "fat"
        ],
        "title": "MealItemResponse",
        "type": "object"
      },
      "MealRequest": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NestedMealItemRequest"
                },
                "maxItems": 500,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "meal_type": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Meal Type",
            "type": "string"
          },
          "time": {
            "default": "",
            "title": "Time",
            "type": "string"
          },
          "total_calories": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Total Calories",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "meal_type"
        ],
        "title": "MealRequest",
        "type": "object"
      },
      "MealResponse": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MealItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "meal_type": {
            "title": "Meal Type",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "string"
          },
          "total_calories": {
            "default": 0,
            "title": "Total Calories",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "date",
          "meal_type",
          "time",
          "total_calories"
        ],
        "title": "MealResponse",
        "type": "object"
      },
...
      "NestedMealItemRequest": {
        "properties": {
          "calories": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Calories",
            "type": "integer"
          },
          "carbs": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Carbs",
            "type": "integer"
          },
          "fat": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Fat",
            "type": "integer"
          },
          "id": {
            "default": 0,
            "minimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "meal_id": {
            "default": 0,
            "minimum": 0.0,
            "title": "Meal Id",
            "type": "integer"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "protein": {
            "default": 0,
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Protein",
            "type": "integer"
          }
        },
        "required": [
          "name"
        ],
        "title": "NestedMealItemRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "MetricsBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "GopherFit API",
    "version": "2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/nutrition/macros": {
      "get": {
        "operationId": "macro_goals_nutrition_macros_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MacroGoalsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No macro targets configured"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Macro Goals",
        "tags": [
          "nutrition"
        ]
      },
      "put": {
        "operationId": "update_macro_goals_nutrition_macros_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MacroGoalsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MacroGoalsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Macro Goals",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/nutrition/meals": {
      "get": {
        "operationId": "meals_nutrition_meals_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 1000000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MealResponse"
                  },
                  "title": "Response Meals Nutrition Meals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Meals",
        "tags": [
          "nutrition"
        ]
      },
      "post": {
        "operationId": "create_meal_nutrition_meals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Meal",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/nutrition/meals/{id}": {
      "delete": {
        "operationId": "delete_meal_nutrition_meals__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "ge": -9223372036854775808,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Meal",
        "tags": [
          "nutrition"
        ]
      },
      "get": {
        "operationId": "meal_nutrition_meals__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "ge": -9223372036854775808,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Meal",
        "tags": [
          "nutrition"
        ]
      },
      "put": {
        "operationId": "update_meal_nutrition_meals__id__put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "ge": -9223372036854775808,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Meal",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/nutrition/meals/{id}/items": {
      "post": {
        "operationId": "create_item_nutrition_meals__id__items_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "ge": -9223372036854775808,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Item",
        "tags": [
          "nutrition"
        ]
      }
    },
    "/nutrition/meals/{id}/items/{itemId}": {
      "delete": {
        "operationId": "delete_item_nutrition_meals__id__items__itemId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "ge": -9223372036854775808,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "ge": -9223372036854775808,
              "le": 9223372036854775807,
              "title": "Itemid",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Item",
        "tags": [
          "nutrition"
        ]
      },
      "put": {
        "operationId": "update_item_nutrition_meals__id__items__itemId__put",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "itemId",
            "required": true,
            "schema": {
              "gt": 0,
              "le": 9223372036854775807,
              "title": "Itemid",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, expired or revoked access token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found or not owned by the caller"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database or stored-data failure"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Item",
        "tags": [
          "nutrition"
        ]
      }
    },
...
  }
}